core: 64-bit tape offsets, 32-bit token lengths, hardened chunk eviction
Multi-GiB IFC support in the parsing core:
- IfcLine.tapeOffset, GetSetArgument/GetSetListArgument, all offset-taking
  loader methods, IfcFileStream::Go and LoaderSettings.MEMORY_LIMIT widened
  from uint32 to 64-bit; variable-length token lengths widened from uint16
  to uint32 (silent truncation of STEP strings >64KiB before)
- SetTokenSource(istream): clear() before seekg — after a read hit EOF,
  failbit made every chunk reload silently return 0 bytes (only reachable
  when the tape exceeds the memory limit, impossible under wasm32)
- IfcFileStream::load(): fetch the exact lookback byte for Prev() on every
  window move; a stale byte made eviction reloads re-tokenize numbers with
  a spurious minus sign or misread '*' as a comment
- comment scanning is EOF-guarded; chunk reloads log when they diverge
- eviction never frees the chunk under the read cursor or writer-fed
  chunks (their content does not exist in the source file)
- GetSchema(): walk STRING tokens and skip payloads correctly (it returned
  IFC2X3 for every file; dead code upstream, load-bearing for the native
  bindings), and don't crash on files without FILE_SCHEMA
- ModelManager::GetSettings no longer returns a reference to a temporary

Verified: 4.7GiB model parses under a 1GiB memory limit with random access
across the whole express-ID range; ctest, jest 90/90 and the 30-fixture
regression suite stay green.

+++ src/cpp/web-ifc/geometry/IfcGeometryLoader.cpp
@@ -235,7 +235,7 @@ namespace webifc::geometry
       std::unordered_map<uint32_t, std::set<double> > mapCurveToDistanceAlong;
       std::vector<glm::dmat4> vecCrossSectionPositionsFallback;
 
-      for ( uint32_t offset : CrossSectionPositionOffsets)
+      for ( uint64_t offset : CrossSectionPositionOffsets)
       {
         auto CrossSectionPositionID = _loader.GetRefArgument(offset);
         // IfcAxis2PlacementLinear
@@ -1501,7 +1501,7 @@ namespace webifc::geometry
     return IfcCurve();
   }
 
-  IfcTrimmingSelect IfcGeometryLoader::GetTrimSelect(uint32_t DIM, std::vector<uint32_t> &tapeOffsets) const
+  IfcTrimmingSelect IfcGeometryLoader::GetTrimSelect(uint32_t DIM, std::vector<uint64_t> &tapeOffsets) const
   {
     IfcTrimmingSelect ts;
